MOB-1338: test changes
